Student's Marks Prediction Using Machine Learning
Education is very important issue regarding development of a country. The main
objective of educational institutions is to provide high quality education to its students.
One way to accomplish this is by predicting student's academic performance and thereby
taking early steps to improve student's performance and teaching quality.
2.The ultimate goal of any educational institution is oering the best educational
experience and knowledge to the students. Identifying the students who need extra
support and taking the appropriate actions to enhance their performance plays an
important role in achieving that goal.